Tower functionality update
I added the functionality to the upgrade and sell buttons on the towers and adjusted currency to change when the events happen. I also fixed a couple of bugs inluding currency not displaying every time it changed and towers sticking to mouse if they cannot be placed or stacking on top of one another when spamming the button.