default tower update
I replaced the sprite of the default tower on the button and prefab(what is placed on the map) to the earth tower in the assets folder. I made sure to fix the firing position but didn't change the tower to the projectile. I also didn't make the upgrade stars look nice since I didn't want to mess up the code for those 2 scripts.