Added the ability for explosive for splash tower
I made it so the splash tower shoots an explosive constantly at a specific spot of the screen, hitting every enemy in its range with fire damage.  The location can be changed from the towerUI panel.  I also changed it so that the splash tower does fire damage with its laser on its second upgrade.