Added Animations and other visual changes
I made it so that the explosive won't fire at the location of the mouse when moving the target.

I also added new sprites for the crosshairs and the coins.

I also added animations to the explosive, the fire damage icon, and the cows while walking, with new sprite sheets for each one respectively.